+++ src/app/tools/page.tsx
@@ -36,7 +36,7 @@ export default function ToolsPage() {
           <div className="mt-4 grid grid-cols-1 gap-2 md:grid-cols-2 lg:grid-cols-3">
             <Link
               href="/tool/keyword-research/overview"
-              className={`group flex h-full flex-row items-center rounded-md border-2 border-slate-200 text-lg font-medium transition hover:bg-slate-50`}
+              className={`group flex h-full flex-row items-center rounded-md border-2 border-slate-200 text-lg font-medium transition hover:bg-slate-50 active:scale-95 active:duration-75`}
             >
               <div
                 className={`flex h-full items-center justify-center px-4 md:px-5`}
@@ -61,7 +61,7 @@ export default function ToolsPage() {
             </Link>
             <Link
               href="/tool/keyword-research/suggestions"
-              className={`group flex h-full flex-row items-center rounded-md border-2 border-slate-200 text-lg font-medium transition hover:bg-slate-50`}
+              className={`group flex h-full flex-row items-center rounded-md border-2 border-slate-200 text-lg font-medium transition hover:bg-slate-50 active:scale-95 active:duration-75`}
             >
               <div
                 className={`flex h-full items-center justify-center px-4 md:px-5`}
@@ -83,7 +83,7 @@ export default function ToolsPage() {
             </Link>
             <Link
               href="/tool/keyword-research/autocomplete"
-              className={`group relative flex h-full flex-row items-center rounded-md border-2 border-slate-200 text-lg font-medium transition hover:bg-slate-50`}
+              className={`group relative flex h-full flex-row items-center rounded-md border-2 border-slate-200 text-lg font-medium transition hover:bg-slate-50 active:scale-95 active:duration-75`}
             >
               <span className="absolute top-0 right-0 rounded-bl-md bg-sky-950/10 px-2 py-1 text-sm leading-none font-medium text-sky-950">
                 FREE
@@ -119,7 +119,7 @@ export default function ToolsPage() {
           <div className="mt-4 grid grid-cols-1 gap-2 md:grid-cols-2 lg:grid-cols-3">
             <Link
               href="/tool/competitive-research/overview"
-              className={`group flex h-full flex-row items-center rounded-md border-2 border-slate-200 text-lg font-medium transition hover:bg-slate-50`}
+              className={`group flex h-full flex-row items-center rounded-md border-2 border-slate-200 text-lg font-medium transition hover:bg-slate-50 active:scale-95 active:duration-75`}
             >
               <div
                 className={`flex h-full items-center justify-center px-4 md:px-5`}
@@ -144,7 +144,7 @@ export default function ToolsPage() {
             </Link>
             <Link
               href="/tool/competitive-research/keywords"
-              className={`group flex h-full flex-row items-center rounded-md border-2 border-slate-200 text-lg font-medium transition hover:bg-slate-50`}
+              className={`group flex h-full flex-row items-center rounded-md border-2 border-slate-200 text-lg font-medium transition hover:bg-slate-50 active:scale-95 active:duration-75`}
             >
               <div
                 className={`flex h-full items-center justify-center px-4 md:px-5`}