+++ src/components/Sidebar/index.tsx
@@ -5,16 +5,22 @@ import {
   BookOpenTextIcon,
   LoaderPinwheelIcon,
   MenuIcon,
+  SettingsIcon,
   TelescopeIcon,
   TextSearchIcon,
+  ToolCaseIcon,
   XIcon,
 } from "lucide-react";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import { memo, useState } from "react";
+import DFSBalanceBox from "../DFSBalanceBox";
+import { Tooltip } from "@heroui/react";
+import Image from "next/image";
 
 const Sidebar = () => {
   const pathName = usePathname();
+  const isSettingsPageActive = pathName === `/account/settings`;
   const [responsiveOpen, setResponsiveOpen] = useState<boolean>(false);
 
   const isToolActive = (slug: string) => {
@@ -24,13 +30,13 @@ const Sidebar = () => {
   return (
     <>
       <button
-        className="absolute top-0 left-0 z-50 flex cursor-pointer rounded-br-md border-r-2 border-b-2 border-slate-200 p-2 lg:hidden"
+        className="absolute top-0 left-0 z-50 flex cursor-pointer rounded-br-md border-r-2 border-b-2 border-slate-200 p-2 active:scale-95 active:duration-75 md:right-0 md:left-auto md:rounded-br-none md:rounded-bl-md md:border-r-0 md:border-l-2 lg:hidden"
         onClick={() => setResponsiveOpen(!responsiveOpen)}
       >
-        {responsiveOpen ? <XIcon size={22} /> : <MenuIcon size={22} />}
+        {responsiveOpen ? <XIcon size={20} /> : <MenuIcon size={20} />}
       </button>
       <div
-        className={`seotoolsuite-sidebar absolute z-50 hidden w-full shrink-0 flex-col gap-3 overflow-y-auto border-b-2 border-slate-200 bg-white py-4 opacity-0 lg:relative lg:flex lg:max-w-60 lg:border-r-2 lg:border-b-0 lg:opacity-100 starting:opacity-0 lg:starting:opacity-100 ${responsiveOpen ? "flex! opacity-100" : ""}`}
+        className={`seotoolsuite-sidebar absolute z-50 hidden min-h-[calc(100vh-62px)] w-full shrink-0 flex-col gap-3 overflow-y-auto border-b-2 border-slate-200 bg-white py-4 opacity-0 lg:relative lg:flex lg:min-h-auto lg:max-w-60 lg:border-r-2 lg:border-b-0 lg:opacity-100 starting:opacity-0 lg:starting:opacity-100 ${responsiveOpen ? "flex! opacity-100" : ""}`}
         style={{
           transition: "all 0.3s ease allow-discrete",
         }}
@@ -97,6 +103,48 @@ const Sidebar = () => {
               <TextSearchIcon size={18} /> <span>Keywords</span>
             </Link>
           </div>
+        </div>
+        <div className="mt-auto h-0.5 w-full bg-slate-200 lg:hidden"></div>
+        <div className="flex flex-row flex-wrap items-stretch gap-2 px-2 lg:hidden">
+          <DFSBalanceBox />
+
+          {pathName.startsWith("/account/") && (
+            <>
+              <Tooltip content="SEO Tools">
+                <Link
+                  href="/tools"
+                  className={`flex items-center gap-1 rounded-md border-2 border-slate-200 px-2 py-2 text-black/80 transition hover:bg-slate-100 active:scale-95 active:duration-75`}
+                >
+                  <ToolCaseIcon size={24} />
+                </Link>
+              </Tooltip>
+            </>
+          )}
+          <Tooltip content="Settings">
+            <Link
+              href="/account/settings"
+              className={`flex items-center gap-1 rounded-md border-2 border-slate-200 px-2 py-2 text-black/80 transition hover:bg-slate-100 active:scale-95 active:duration-75 ${isSettingsPageActive ? "bg-slate-100!" : ""}`}
+            >
+              <SettingsIcon size={24} />
+            </Link>
+          </Tooltip>
+          <Tooltip content="Discord Server">
+            <Link
+              href="https://discord.gg/Wt4RN4Xy8n"
+              rel="nofollow"
+              target="_blank"
+              className={`flex items-center gap-1 rounded-md border-2 border-slate-200 px-2 py-2 text-black/80 transition hover:bg-slate-100 active:scale-95 active:duration-75`}
+            >
+              <Image
+                src="/assets/images/discord-icon.svg"
+                width={20}
+                height={20}
+                quality={100}
+                alt="Discord"
+                className="w-5"
+              />
+            </Link>
+          </Tooltip>
         </div>
       </div>
     </>